refactor(engine): use immutable zeebe state
The processors are not allowed to change state. Generally they are
provided the immutable variant of the ZeebeState interface. This class
was still using the MutableZeebeState because it used it to gain access
to the key generator (the only exception of the mutable states).

This commit removes the usage of the MutableZeebeState from the
ProcessEventProcessors to make it easier to provide the immutable
states to the processor constructors.