test(engine): should activate root element
Adds a failing test that shows that the process instance modification
with activate instruction can activate a root element in the process.

It does so by creating a process instance which waits at a service task,
and modifies the process instance by adding an additional token, waiting
at the other service task.

It's important that this test verifies the details in the
ELEMENT_ACTIVATED event produced as a result of processing the
modification commmand, because this is the public api.