feat(backup): checkpoint listeners are notified after init and replay
The listeners must be updated with the current checkpointId after init and replay. Otherwise if there are no new checkpoint created, after a restart or failover the listeners cannot know about the latest checkpoint.
It is important for `InterPartitionCommandSender` and `InterPartitionCommandReceiver` to keep track of the latest checkpointId to guarantee that the checkpoints are consistent.