test(engine): skip init for same input/output vars
When the output element expression is a variable (or a nested property
of a variable) then it needs to be nil initialized, to make sure that
the variable exists at the local scope.

However, when the output element expression refers to the same variable
as the input element, then it shouldn't overwrite this variable. Note
that the input variable is already created at the local scope.

This adds a test case that fails if the output element variable is nil
initialized when it refers to the same variable as the input element.