fix(engine): dont overwrite local input element var
When the output element expression is a variable (or a nested property
of a variable) then it needs to be nil initialized, to make sure that
the variable exists at the local scope.

However, when the output element expression refers to the same variable
as the input element, then it shouldn't overwrite this variable. Note
that the input variable is already created at the local scope.

We can simply filter out this case, and only nil init the other output
element variables.

(cherry picked from commit 1377cd64a0c23373e814a5bddce1171b7846bdd5)