test(engine): skip init for loopCounter as output
When the output element expression is a variable (or a nested property
of a variable) then it needs to be nil initialized, to make sure that
the variable exists at the local scope.

However, when the output element expression refers to the loopCounter
variable, then it shouldn't overwrite this variable. Note
that the loopCounter is already created at the local scope.

This adds a test case that fails if the output element variable is nil
initialized when it refers to the loopCounter variable.