feat(broker): abort snapshot that is concurrent to a backup
This is required to prevent inconsistencies in the backup that is triggered by the checkpoint. A backup must always include a snapshot at position before the checkpoint position. If the checkpoint is taken while the snapshot is also taken, then it is possible that the actual snapshot position in the snapshot is greater than the checkpoint position because the snapshots are taken asynchrously. If we include this snapshot in the backup, the backup is not consistent anymore. To be safe, we abort the snapshot. This is not an ideal solution, but it is a simple solution to prevent the inconsistency.