feat: modify PI by activating element inside existing flow scopes
Allows the activation of nested elements, as long as they are activated inside of an existing flow scope. Using the flowscope element id we can search for an element instance matching this id. When found we can return the key of this element instance and use it as the flow scope when activating the element from the activate instruction.