build(docker): add OCI/OpenShift labels to Docker image
Adds labels to the Docker image following the OCI and OpenShift specs.
This includes modifications to the build process to inject the few
values which are dynamic, namely:
  - the created at ISO 8601 timestamp
  - the commit SHA (or revision) of the artifact
  - the semantic version of the artifact

On top of adding labels, this modifies the `Dockerfile` a bit and pins
the production image to specific sha of the base image, ensuring
reproducible builds. In the future we should update this sha when need
be, and update the golden file
(`docker/test/docker-labels.golden.json`).

This also adds `hadolint` to lint our Dockerfile and applies some of the
recommendations to it. A new code quality job is added, `Docker checks`,
which runs hadolint and verifies that the labels are as expected.