refactor(engine): clarify error message constants
These constant names are not fitting any more. They are the error
messages that are logged on specific situations, but the engine has
changed so much that these names no longer match the case.

The first error message is about not being able to find the related
record processor.

The second error message is about processing a record but an exception
occurred.