refactor(engine): explicitly handle unexpected error
So far, all processing errors have been unexpected. In that case we log
an error with lots of metadata about the record, write a command
rejection, and potentially write an Error event that blacklists the
related process instance.

Now, we want to support expected errors in the same place. We'll need to
differentiate between these two types of processing errors. We can do
that with the ProcessingError enum, which names the two modes. I
preferred this over a simple boolean value.

When a processing error is expected then we only want to return the
results we've added to the result builder. Note that this result will
probably contain a specific rejection (and potentially some follow-up
records), but probably no Error record. In the end this is up to the
respective TypedRecordProcessor. We'll introduce this in an upcoming
commit.