test(engine): reject PI creation atomically
When we can't subscribe to the catch events as part of the Start Process
Instance Anywhere feature of ProcessInstanceCreation, then we should
reject the command *and* not write any other events. I.e. the process
instance should not be activated and no element instances should be
created.

We can test this by making sure there are no PROCESS_INSTANCE
ACTIVATING/ACTIVATED events on the log after the
ProcessInstanceCreation CREATE is rejected.