refactor(engine): allow processor to handle error
If we can find the relevant record processor for the failed command,
then we make that processor try to handle the error.

By default all errors are treated as UNEXPECTED. If a processor
overrides the tryHandleError method than it can choose which errors to
EXPECT and which not by returning a different value.

If we want to be able to use this mechanism in the
CreateProcessInstanceProcessor, then the CommandProcessor (which wraps
some of the processors) should also support this.