fix(streamprocessor): reset writers on error
Originally, the writers were reset in the processing state machine
(using the no longer available method `resetOutput`), when an error was
detected. We assume that this got lost in the refactoring.

By resetting the ProcessingResultBuilder, we achieve the same thing
(resetting the writers).

Note that the ProcessingResultBuilder is constructed just before, which
gives the false feeling that the result builder is complete new and
fresh, but that's not entirely true. A call to reset is still necessary.