fix(streamprocessor): don't log caught errors
The engine explicitly handles errors caught by the stream processor
during command processing. There's no need for the stream processor to
also log these errors.

This has become especially important now that the engine starts to throw
exceptions for expected errors. Such errors should not be logged at the
ERROR severity because they should not be reported in Google's Error
Reporting Tool.

Therefore, we should remove this logging from the stream processor, and
allow the engine to decide when to log an error as ERROR severity and
when to handle it as an expected error.