feat(grafana): clarify DEAD partition state
A partition can either be HEALTHY, UNHEALTHY or DEAD. However, we never
had any way to show this in the partition health overview.

This adds an additional field override for health -1 -> DEAD with the
red color. To differentiate it from UNHEALTHY, I've also changed the
color of UNHEALTHY to orange.