docs(engine): clarify expected error
The EventSubscriptionException is private to this class, and is only
thrown during flow scope activations for start instructions.

E.g. when the start instruction wants to activate a task inside a
subprocess with a boundary event. If we can't subscribe to the boundary
event (e.g. missing variable), then the exception is thrown.

The exception is not thrown during regular process instance processing.