fix(engine): lookup pending distro before deleting
The pending deployment is deleted when the DeploymentDistribution
COMPLETED event is written.

We should only write this event if the pending deployment still exists.
If it doesn't exist it either was already completed or it was never
pending. In both cases, the command to COMPLETE the deployment
distribution should be rejected.

We can achieve this by providing a way to check whether a specific
deployment distribution exists in the state. The current method only
allows to check that there is at least one pending distribution for a
specific deployment, but it doesn't allow to check the specific
partition id.

(cherry picked from commit 6db151ecf7578ff97c6aa7b590614fc81d666293)