test: fix flaky MultiPartitionDeploymentLifecycleTest
Test was flaky because it was asserted that the COMPLETE commands and COMPLETED events were written in order.
This is not always as the case. The COMPLETE commands are a result of partition 2 and 3 sending the command to partition 1. There is no guarantee that partition 1 will first receive both commands before it starts processing them.
As long as we assert that we receive the COMPLETE command of partition X before the COMPLETED event of partition X the test should be accurate without being flaky.