refactor(engine): migrate timeout trigger to task with record result
It is not a pure refactoring:

Previously the code checked whether the flush succeeded.
The new code assumes that a later flush will succeed.

This is a change and behavior beyond a simple refactoring.

This commit also removes JobTimeoutTriggerTest. This test had just one testcase.

Also adds method to ProcessingScheduleService to
schedule a task with a delay.