fix(test-util): apply compact logger abbrv in order
The abbreviations of the compact logger are applied on the text
sequentially. So far the order didn't matter because the keys don't
overlap. But, I want to add some new keys that do overlap.

This changes the abbreviations from a Map to a List of Entries to
strictly preserve the defined ordering.

(cherry picked from commit 9590133120ab3237d04fca38438f27a80f540d8f)