feat(test-util): abbreviate dply and dstr records
Deployment and Deployment_Distribution record types should be
abbreviated correctly. However, only DISTRIBUTED was abbreviated, which
is an intent and not a record type. This led to some strange logs when
reading the compact log.

By removing the DISTRIBUTED abbreviation and adding abbreviations for
the record types we get a much cleaner log.