test(qa): only check that the topology reports all replicas
Instead of using the helper `isComplete`, manually check that the
topology contains all replicas, omitting the check for the partition
health. While this would be safer, there is a race condition where the
partitions may be unhealthy for a short time on start up. For this test,
we only really want to check that the nodes find each other, not that
processing works, as it should be largely unaffected.