test: verify client response
The client should get a describing message as a response when the modify command gets rejected.

Since the command requires a valid key to resolve the partition it belongs to, I use the processDefinitionKey here. This way it won't resolve to a process instance, but it's still valid and we can extract the partition from it.