build(.github): disable setup-maven on macOS
Removes the `maven-version` input for the `setup-zeebe` action, and
instead hard codes the version (since we want the same version
everywhere, and there's no use case to customize it yet). Instead, adds
a new input `maven`, which is a boolean param, controlling whether maven
will be installed or not.

Hopefully this fixes the issues on macOS where setting up maven times
out, or we run out of cache code space.