test(qa): ensure long polling activates jobs
Refactors the LongPollingIT to be of narrower scope. The main idea is
that the test now does the following:

- The first activate command is processed before the create process instance command
- Once the job created event is applied, a new job activate command should be sent again without the client having done so
- The client receives the job

To do that, we send the activate requests manually (to avoid the worker
re-polling and thus getting the job without going through long polling).
This ensures the client only sent a single request, and that if there
are two ACTIVATE commands on the logs, it must have come from long
polling.

Refactoring also includes:

- Simplification of set up (e.g. dumping logs of all nodes in a single
  extension, only using a single partition, etc.)
- Setting the long polling logger to trace for more information on
  failure
- Using `ContainerEngine` to be able to assert the ordering of the
  commands on the log, as well as getting the log printed out on failure