test(qa): fix flaky assertion
Since the notification for jobs available is sent on state update, we
cannot guarantee the ordering of the subsequent ACTIVATE command with
the process instance records. It's anyway not strictly required for us
to wait for the job to exist, it's enough to check that a second command
is sent without the client having to send one explicitly.