feat(backup): coordination logic in BackupService
For testing, we split the actor part and the actual implemenation of the BackupService. BackupServiceImpl implements the actual logic for coordinating taking backup of a single partition. For testability, we also added a InProgressBackup. This commit only includes the implementation on BackupServiceImpl. Only after implementing InProgressBackup, a backup will be taken.

Since the nesting calls on ActorFuture is no very readable, in BackupServiceImpl we implemented a naive way of composing the futures. If a step failed, the corresponding future is failed and futures for all the remaining steps are failed transitively, finally failing the `backupSaved` future.