feat(engine): apply global var instructions
The variable instructions for global variables should be applied before
the target element is activated and before its element scopes are
created.

Locally scoped variable instructions will need to be executed as part of
the activation / flowscope-creation logic in the
ElementActivationBehavior, because each variable instruction requires a
flow scope key (that is looked-up using the variable scope id). This
means that the activation behavior will become intertwined with variable
instruction logic.

It may be that the global variable instructions can be done elegantly
inside this same logic, but I don't yet know what this logic will look
like for local variables. So, for now, I've placed the entire logic for
global variable instructions outside of the ElementActivationBehavior.

When we look at local variable instuctions, we should spend some time to
see if we can merge these parts and remove the duplication. But if this
leads to additionally complex code, it's likely better to just duplicate
the logic.

Also, please note that an unexpected exception is thrown if the variable
instruction is not an instance of our own implementation. This is
necessary to retrieve the variables direct buffer. Otherwise, we can
still retrieve the variables as a Map<String,Object>, but serializing
this into MsgPack wasn't easy. I therefore decided to throw an
exception, since this is not a case we want to support anyways.