feat(engine): reject modify on unknown activate target
If any of the activate instructions of the ProcessInstanceModification
command refer to an element id for which the element cannot be found in
the relevant process, then the command must be rejected.

This could be done while processing each of the activate instructions,
but that would require throwing an exception to roll back any already
executed instructions. Instead, this implementation first checks whether
there are any elements that can't be found. If so, it rejects before it
even started to execute the instructions.

This also allows to collect all the missing element ids and inform the
user at once. If we would only provide the first we encountered, then
the user could go through a series of modify commands before resolving
all the argument-related problems. If they are provided all these
problems at once, its much more clear for them.

I've also chosen to add the process ID to the rejection reason to
clarify which process is actually looked at. We could argue that the
process version or process key (or even the deployment key) could also
be useful, but I wasn't sure about it. I think this is enough
information for an initial investigation by the user.