refactor(engine): extract modify command validation
This moves the command validation to a separate method, to make it
easier to add more validations that can be performed before executing
the instructions.

We could consider to collect multiple validations, but there's no good
way to reject the command for multiple reasons. A rejection has a single
rejection type and a single rejection reason. The reason can be enhanced
to contain multiple reasons, because its a simple string, but that
doesn't work when there are multiple rejection types. Since this seems
to become complex, I'd rather keep it simple and return a reason as soon
as we find it. Therefore, this method return Optional<Rejection>.

Note that I intentionally return Optional.empty() at the end of the
method, so that it can stay the end of the method even when we add more
validations to it.