feat(bpmn-model): accept duration type for timer
A timer with duration can only be specified as a static value using a
String (in the ISO 8601 format), or as a FEEL expression. However, for
the Java bpmn model API it makes a lot of sense to pass a
java.time.Duration instance.

This makes tests a bit easier to create because you no longer have to
consider what this ISO format looks like again. You can use the commonly
used java.time.Duration instead.

For example: `timerWithDuration(Duration.ofMinutes(10))`