test(engine): reject activate event-based-gw event
This adds the general test infra to verify modify command rejections due
to unsupported element type. The goal of this test class is to
eventually be entirely removed, because the we only unsupport particular
types for technical reasons.

This also introduces the first scenario for such a rejection: events
belonging to an event-based gateway as the target of an activate
instruction.

This test class is setup in such a way that it should be easy to extend
it with more scenarios.