fix(engine): reject unknown element as invalid_arg
This changes the rejection type of the ProcessInstanceModification
command with activation of an unknown element from NOT_FOUND to
INVALID_ARGUMENT.

We do this to align the behavior with the ProcessInstanceCreation
command where no element can be found for a start instruction.

In CreateProcessInstanceProcessor, we reject the command with the type
INVALID_ARGUMENT if one element id of an activation instruction is
unknown. We use NOT_FOUND only if the target entity (i.e. the process
instance in this case) doesn't exist.