feat: support new canWrite check on writer
We have some use case where we want to verify whether we could potentially write certain event count and event batch size to the LogStreamBatchWriter, without already adding it to the batch.
This commit adds a new method to check whether the event count and batch size could potentially written.