feat(test-util): compactly log record timestamps
When troubleshooting issues or writing tests for Timer related topics,
we're often not only interested in the Duedate of the timer, but also in
the Timestamp of the records.

Since these timestamps are generally irrelevant for other types of
tests, it makes sense to only log them when there are Timer records
present.

As a format I've chosen a small but simple form (HH:mm:ss). If needed we
could investigate more advanced formatting in the future. E.g. a format
that only shows the least amount of data: only seconds if those matter,
or include hours if those matter. For now, this short form in
combination with hiding it when its not useful seems useful and easy to
build.

Timestamps are naturally found at the start of log lines, so I've chosen
to prefix each record summary with it.