fix(test-util): only print partition in legend if needed
If there are multiple partitions, the compact record logger's Legend
should contain [Partition] and each key will contain a [Pn] prefix, so
readers understand the summary.

But, there aren't always multiple partitions, and then the partition
isn't part of the compact log. In that case, the [Partition] should
also be removed from the Legend.

I've also split up the Legend construction a bit further, to keep it
somewhat readable.