feat: introduce canAppend method to RecordBatch
This method can be used to verify whether a record length could potentially appended to the batch, this is useful if you don't want to append the record right now, like if you add more and more data and always check whether it still fits. This is used in the JobBatchCollector for example.