fix(raft): reset snapshot replication on error in follower
When follower responds with any error, leader restarts send the snapshot from the initial chunk. When the follower has not resets its state, then the follower is not expecting the initial chunk. As a result, the follower reject the requests and this will continue endlessly.

(cherry picked from commit 9aba732358778c277557eba97d47d4893ef73dc5)