test(engine): move reject modify for multi-instance
ProcessInstanceModification commands with Activate instructions that
refer to elements inside a multi-instance body are rejected because they
are unsupported. These are unsupported for the same reason as the other
unsupported element types.

In a previous commit, I had created a new test class called
ModifyProcessInstanceUnsupportedElementsTest. The intention of this
class is to eventually be removed when we are technically able to.

This is separate from the ModifyProcessInstanceRejectionTest test class
which covers all cases for command rejections that are part of the
feature itself, e.g. user mistakes like referring to non existing
elements.

Therefore, it makes sense to move the rejection test for multi-instance
into the unsupported elements test.

An additional reason for testing these in the same place is to align the
relevant rejection messaging. Therefore this commit is not just a
refactoring but a test change. After this commit, the test fails because
the verified message does not align yet.