feat(engine): reject modify for event-based-gw event
Events belonging to an event-based gateway are unsupported, because they
cannot technically be activated with an activate instruction at this
time. We should aim to support this in the future.

This commit builds on top of the previous rejections, using a similar
rejection message style. The message is set up for re-use among any of
the unsupported elements.