refactor(backup): find segments files before finding the snapshot
There is potential race conditition if segment files are collected after finding the snapshot:
1. Find snapshot -> no snapshot available
2. Do not reserve snapshot because no snapshot available
3. New snapshot is taken and segments compacted
Collect segments -> but initial segments are deleted. As a result backup will be incomplete

To prevent this case, we first collect the segment files. The records until the checkpoint position is guaranteed to be in the segment at this point. Now if a snapshot is taken and the segments are deleted, the backup will fail because backup store cannot read those files. This is not optimal, but still prevents unknowingly taking an incomplete backup.