feat(engine): reject modify activate start event
ProcessInstanceModification command with activate instructions
targetting elements of type start event are not supported, due to
technical reasons. We should aim to support these in the future.

This implementation immediately goes a bit further than just start
event, as it allows for more unsupported element types to be easily
added.