feat: reject too large deployment
When a deployment was too large an unexpected exception would be thrown. This would cause a rejection to be written with an error message containing the very large resource. The writing of this rejection failed because it was too large as well, causing another rejection to be written. This kept happening in a loop.

By manually writing the rejection in the processor we can give a custom message that isn't too large, resolving the loop.