test(engine): reject modify activate unsupported elements
When multiple elements are referenced in the activate instructions of a
ProcessInstanceModification command, then the rejection message should
collect both the element ids as well as the element types of the
culprits.

By highlighting several of the same class of rejections, users can
resolve the type of problem faster, without having to retry the
modification command.