fix(engine): unflake trigger only twice
This test became flaky after a change in the task scheduler. Since then,
tasks that have been scheduled are only executed when the engine is in
the PROCESSING phase, but isn't currently processing a command. If that
is the case, then it is rescheduled. That means that sometimes things
take longer, especially when the processor is busy.

In this specific test case, new process instances are created using 2
timer start events on different processes. Each time they trigger, they
produce a bunch of records because they start new process instances.
The processing of the relevant commands for these instances can
interfere with the DueDateTimerChecker which is a scheduled task. Because
of the new task scheduler change, it can happen that the
DueDateTimerChecker is postponed.

When that happens, it becomes a responsibility of the test case to make
sure that the clock is moved further again to make sure that this
rescheduled task is actually run.

Potentially this is less of a problem when the clock is unpinned, and
the actual system time is used. But, IMO this also highlights the
fragility of these tests. I prefer control in the tests (with a pinned
clock), even if we have to work around some specific quircks of our
system.

By awaiting the end of the log, we know we can increase the time and
reliably expect the DueDateTimerChecker to be run.