test(engine): control the clock in timer tests
We should take full control over the clock in tests that verify the
correct workings of our Timer events. With the TimerStartEventTest being
flaky, let's add some control by pinning the clock's current time. This
makes sure the clock no longer moves on it's own.

Note that this increases the flakiness of the shouldTriggerOnlyTwice
test, but we can fix that in an upcoming commit.