refactor: support multiple variable instruction for one activate instruction
It's possible for an activate instruction to contain multiple variable instructions. This is to enable users to add variables in different scopes. By using a builder-like structure new variable instruction can be appended.

To achieve this I've created a new inner class. Upon adding an activate instruction this class is returned. The relevant activate instruction is passed so we know on which activation we should add the variables.

Finally, in order to make this work we can't add the activate instruction to the record immediately when calling the activateElement method. Adding an instruction to the record creates a copy of the instruction. If we were to add it immediately it'd no longer be possible to add variable instructions. To circumvent this a list of activate instructions is being stored. Once the modify method is called all activate instructions will be added to the record.