test(engine): reduce test dependency
I was thinking that awaiting the engine to be idle before increasing the
time, is not a great way to run a test. It's too specific.

Instead, we can just use what we have available: the recording that
shows how often a timer has been triggered so far. If it's not yet 5, we
simply move the clock ever so slightly.

However, it's not that easy, because we have to provide an expectation
over the number of expected records (limit). And we await for up-to 5s
before making the assertion. So, if it only triggered 4 times at the
moment we check it, then we have to wait 5 seconds until we can increase
the time.

So this test has become a lot slower whenever it misses the
DueDateTimerChecker's scheduled task.