feat(test-util): add method to disable awaiting records
Sometimes it's useful to not await until all the records are recorded,
but simply to supply what is currently available. For example, when we
want to increase the clock periodically while we await for a specific
stream of records.

Currently, the awaiting behavior of the RecordingExporter blocks such
logic, because it blocks the call until it is done (i.e. it's
synchronous). We must wait out the blocked call, before we can take
other actions again (like increasing the clock's time).

Using awaitility, we can easily periodically check for a specific
condition. And if it's not yet met, we can execute other code.

So all we need to do is disable the awaiting behavior and combine it
with Awaitility. It was unclear to me that the wait time could be
changed, and how I could disable it. That's why I first built my own
ImpatientRecordIterator to replace the AwaitingRecordIterator. Later I
discovered that I could just set the wait time to 0.

Therefore, I've introduced a new shorthand method to clarify how to
disable the awaiting behavior.

This also explains why I've changed the RecordIterator's name to
AwaitingRecordIterator (because that's what it is).