test(engine): further unflake triggerOnlyTwice
An interesting question was raised: why would we only need to loop for
the last assertion? There's no clear answer for this, apart from its
less likely that the scheduler will postpone the DueDateTimerChecker
task for a busy engine at that moment.

However, of course we should make sure there is no such situation in the
first place. So, we decided to also loop on the second assertion.