perf: don't validate duplicates
Each instruction is individually checked for specific requirements. The
results are collected in a Set, so they are nicely deduplicated in the
rejection message. However, if there are duplicates, we only have to
test them once per validation.

By using distinct early in the collection pipeline, we can deduplicate
the collection pipeline before performing the actual validation. This
can especially be useful when doing costly operations like state
lookups.

Additionally, we can switch to List over Set to maintain the ordering of
the instructions as provided.