refactor(backup/s3): type checked manifest per backup status
This is (another) larger refactoring to allow type checked handling of
manifests. All backup statuses are represented by dedicated manifest
types. Only valid state transitions are allowed. For example, there's
no method that allows marking a completed backup as in-progress again.
Manifest of backups that aren't failed don't have to contain a field for
the failure reason while manifests of failed backups always have a
failure reason.

Manifest type is either `NoBackupManifest`, which can't be serialized
and is only used as a return type when the backup doesn't exist, or
`ValidBackupManifest` which can be serialized and written as an object.
Deserialization is driven by the status code field.