fix(streamprocessor): stream processor can read checkpoint record values
StreamProcessor uses this Event registry to read the value based on the given ValueType.

This event registry and `RecordValues` that reads the record is now part of engine. It is not ideal to add checkpoint value types in the engine, as the engine shouldn't care about it. Alternatively, I tried to move `RecordValues` to StreamProcessor. The goal was that each RecordProcessor can define its own registry which we can feed it into `RecordValues`. However, this did not work because engine uses `RecordValues` in their tests. This will create a dependency from engine to streamprocessor which is not allowed. Engine should only depend on the common api. Besides, ExportedDirector also used `RecordValues` to read the record.

Because of these dependencies, changing anything here would required a lot of refactorings. So as a quick fix, I have added the checkpoint valuetype to the registry. It does not add any additional dependencies to the engine.