fix(protocol-impl): clone var buffer on copy
Copying ProcessInstanceModificationVariableInstructions was broken when
multiple such instructions exist or when multiple
ProcessInstanceModificationActivateInstructions with a variable
instruction exist.

The reference to the buffer was simply set, but when the buffer is
reused for another object, then the pointer of the first still refers to
the now updated buffer.

By cloning the buffer as part of the copy, we make sure the object is
deeply cloned.