feat(gateway): coordinate sending backup request to all partitions and collect responses
Backup request is processed by the StreamProcessor and thus the response will be a ExecuteCommandResponse similar to other streamprocessor commands.

The request handler sends the command to all partition and wait for response from all of them. The backup request is successfull only if all partitions sends a sucess response. When the request is successul, backup has started on all partitions but not necessarily completed.