fix(scheduler): prevent race condition due to submit
By using `ActorControl#submit` in some of the retry strategies, we can
create race conditions if the retry strategy is reused. Since the
initial call uses run to prepend a retry attempt, and further retries
use submit, it's possible for one run to retry (thus submitting the
retry job to the end of the queue) and the next call to `runWithRetry`
cause its state to be overwritten, causing issues when it comes to
completing the future (as well as potential shared state by the
operations).