build(.github): auto merge dependency patch updates
Adds a workflow which runs whenever a PR is opened, reopened, marked as
ready for review (so going from draft to ready), and updated
(synchronized). The latter is to allow things such as `@dependabot
rebase` to trigger this workflow.

This workflow runs only if the author of the PR is dependabot, and if
the repository is `camunda/zeebe`. It will only merge patch updates for
now, but it could be updated to merge any kind of updates.