test: verify termination and activation of element in same scope
When we terminate an element we will terminate the flow scope if this is the only active element in it. If we have an activate instruction for an element in the same scope, we should not terminate the flow scope, as it is still required.

This test will verify that the activated element has the same flow scope as the terminated element.