feat: terminate and activate element in the same scope
When we terminate an element we will terminate the flow scope if this is the only active element in it. If we have an activate instruction for an element in the same scope, we should not terminate the flow scope, as it is still required.

This commit makes it so that when we activate an element using the ElementActivationBehavior we return a new object with the element instance key of the activated element, and a set of all it's flow scopes.
This set of flow scopes is passed to the termination instruction logic. We use this to verify if the scope that's being terminated is required for one of the activate instructions. If it is, we cannot terminate it.