feat(broker): validate s3 config by creating a throwaway S3BackupStore at startup
If configuration like region, endpoint and credentials are not configured in zeebe config and not available from the environement, creating S3BackupStore fails. S3BackupStore is created during role transitions. Then it will fail the transition, and the partition goes unhealthy. To fail fast, when the configurations are not available, we create a throw away S3BackupStore while validating the configs.